feat : add comment and rm unnecessary file

+++ front-bank/src/components/TableOperations.vue
@@ -1,7 +1,7 @@
 <template>
     <div>
         <h3 class="text-center">Historique des opération</h3>
-        <!-- Ici on affiche l'historique de nos opération -->
+        <!-- Here we display the history of our operations -->
         <table class="table table-bordered">
             <thead>
                 <tr>
@@ -18,7 +18,7 @@
             </thead>
             <tbody>
 
-                <!-- Pour aider à la lecture on utile une opération ternaire -->
+                <!-- To help the reading we use a ternary operation -->
                 <tr v-for="submission in formSubmissions" :key="submission.id"
                     :class="submission.type === 'debit' ? 'table-danger' : 'table-success'">
                     <td><i :class="submission.type === 'debit' ? 'fa-solid fa-minus' : 'fa-solid fa-plus'"></i></td>
@@ -46,7 +46,7 @@ export default {
         }
     },
     mounted() {
-        // Récupère les transactions sur l'API Flask
+        // Retrieves operations from the Flask API
         fetch('http://127.0.0.1:5000/last_10_operations')
             .then(response => {
                 return response.json();
@@ -54,7 +54,7 @@ export default {
             .then(transactions => {
                 console.log(transactions)
 
-                // Ajout de chaque transaction dans notre tableau
+                // Adding each operations to our table
                 transactions.forEach((item) => {
                     this.formSubmissions.push(item);
                 });
@@ -64,33 +64,40 @@ export default {
             });
     },
     methods: {
-        // Ajout d'une nouvelle transaction avec le formulaire
-        // On ajoute la nouvelle transaction au début du tableau pour la faire
+        // Adding a new operation in the form
+        // We add the new transaction at the beginning of the table
         addFormSubmission(formData) {
             this.formSubmissions.unshift(formData);
         },
 
+        // Deleting an operation from the history
         deleteFormSubmission(id) {
 
-            fetch('http://127.0.0.1:5000/delete_operations/' + id, {
-                method: 'DELETE',
-                body: id
-            })
-                .then(response => {
-                    if (response.ok) {
-                        const index = this.formSubmissions.findIndex((submission) => submission.id === id);
-                        if (index !== -1) {
-                            this.formSubmissions.splice(index, 1);
-                        }
-                        return response.json();
-                    } else {
-                        throw new Error('Erreur lors de la suppression des données');
-                    }
+            // The deletion is triggered by a click on the trash icon
+            // To avoid errors we ask for a confirmation 
+            const confirmation = prompt("Si vous souhaitez vraiment supprimer cette opération entré \"supprimer\":");
+            if (confirmation === "supprimer") {
+                // Call to the API to delete the identified operation
+                fetch('http://127.0.0.1:5000/delete_operations/' + id, {
+                    method: 'DELETE',
+                    body: id
                 })
-                .catch(error => {
-                    console.log(error);
-                });
-
+                    // If the api answer is successfull, then we delete the line locally too
+                    .then(response => {
+                        if (response.ok) {
+                            const index = this.formSubmissions.findIndex((submission) => submission.id === id);
+                            if (index !== -1) {
+                                this.formSubmissions.splice(index, 1);
+                            }
+                            return response.json();
+                        } else {
+                            throw new Error('Erreur lors de la suppression des données');
+                        }
+                    })
+                    .catch(error => {
+                        console.log(error);
+                    });
+            }
 
         }
     }
